+++ packages/ubcli/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - `ubcli meta-tr` will show line and column in case of invalid meta file JSON
 
 ## [5.5.0] - 2020-02-29
 ### Changed